// src/interfaces/i-landingpage.ts
export interface ILandingPageResponse {
    createdAt: string;
    _id: string;
    homeVideo: string;
    homeTitle: string;
    homeDesc: string;
    serviceVideo: string;
    serviceDesc: string;
    productDesc: string;
    successDesc: string;
    successImage: string;
    successFinishProject: string;
    successHappyCustomer: string;
    successIssueResolved: string;
    successAwardes: string;
    expertTeamDesc: string;
    outcomeDesc: string;
}

export interface ILandingPagePayload {
    homeVideo: string;
    homeTitle: string;
    homeDesc: string;
    serviceVideo: string;
    serviceDesc: string;
    productDesc: string;
    successDesc: string;
    successImage: string;
    successFinishProject: string;
    successHappyCustomer: string;
    successIssueResolved: string;
    successAwardes: string;
    expertTeamDesc: string;
    outcomeDesc: string;
}